+++ game.vcxproj
@@ -259,8 +259,8 @@
     <ClInclude Include="Source\Core\StdAfx.h" />
     <ClInclude Include="Source\Game\config.h" />
     <ClInclude Include="Source\Game\mygame.h" />
-    <ClInclude Include="Source\Game\Raiden\Background\Background.h" />
-    <ClInclude Include="Source\Game\Raiden\Component.h" />
+    <ClInclude Include="Source\Game\Raiden\Background\Background.hpp" />
+    <ClInclude Include="Source\Game\Raiden\Component.hpp" />
     <ClInclude Include="Source\Library\audio.h" />
     <ClInclude Include="Source\Library\gamecore.h" />
     <ClInclude Include="Source\Library\gameutil.h" />
